more reorganization; add `.help()` function for quick reference
