+++ chromatic/rainbows/actions/__init__.py
@@ -1,5 +1,4 @@
 from .binning import *
 from .trim import *
 from .normalization import *
-from .wavelike import *
 from .align_wavelengths import *
